fix[ux]: fix error message for "staticall" typo (#4438)
the "staticall" typo is very common, and was raising a hard-to-read
error message from the python parser like "invalid syntax. Perhaps you
forgot a comma?"

this commit adds logic to `ast/parse.py`, so we can add a hint to the
exception after we have successfully gone through the python parser and
have the ability to add source info and a hint.