fix[parser]: fix bad tokenization of hex strings (#4406)
this commit fixes parsing of hex strings. there were several issues
with the hex string pre-parser, including:

- modification of the string locations
- incorrectly not exiting the state machine if a non-string token is
  encountered.

this commit fixes the state machine, changes the pre-parser to leave
the locations of hex strings unmodified as to minimize the changes to
locations in the reformatted code vs source code. to see the effect,
print out the reformatted code of the test cases included in this PR
before and after this commit.

this commit additionally adds several sanity checks to the pre-parser so
that the chance of future tokenization bugs is minimized.