refactor[lang]: remove VyperNode `__hash__()` and `__eq__()` implementations (#4433)
it is a performance and correctness footgun for `VyperNode`'s hash and
eq implementations to recurse. for instance, two nodes from different
source files should never compare equal.

several tests rely on the recursive behavior of the eq implementation;
a utility function `deepequals()` is added in this PR so that tests
can perform the recursive check on AST nodes where needed. nowhere in
the compiler itself (`vyper/` directory) is the recursive definition
relied on.

this commit also slightly refactors the import analyzer so that uses
the new hash and eq implementations instead of the previous workaround
to avoid recursion (which was to use `id(module_ast)`.