chore[ci]: separate codecov upload into separate job (#4455)
this commit separates `coverage combine` (the local tool) and codecov
upload (upload to 3rd party app) into separate steps. it also uploads
the coverage report artifacts. this makes the actions more transparent
since we can download the coverage database to inspect it if there is
an issue, and allows us to set the `coverage report` as a required CI
check, which does not depend on a 3rd party service. (if it depended on
a 3rd party service, we might then be unable to merge PRs if there is
an issue on the 3rd party service's side).