feat[test]: add hevm harness for venom passes (#4460)
this commit adds `hevm` as a harness to some venom tests. to accomplish
this, it adds a new venom instruction "sink", which in the new
`hevm_check_venom` function gets compiled to a `return` instruction
so that hevm can check the returndata for equivalence. this allows us
to rely on bytecode equivalence of the compiled code to "prove" our
optimizations are sound.

the new pattern has been applied to the tests in
`test_algebraic_binopt.py` as a first pass, but it can (and should) be
applied to more venom tests in the future. and `hevm=False` kwarg has
been added for some cases where hevm cannot resolve the equivalence.

an hevm harness could also be added to vyper sources, but that is out
of scope for this PR and left for future work.

references:
- https://github.com/ethereum/hevm