feat[ux]: improve hint for events kwarg upgrade (#4275)
follow-on commit to ebe3c0ccfc2a6935d. the hint in ebe3c0ccfc2a6935d is
not specific to the user's source code, and it's a bit laborious for the
user to match the call-site to the event def. this commit auto-generates
the new call expression for the user to use. additionally, it migrates
the example contracts in the `examples/` directory so that they stop
emitting warnings.

---------

Co-authored-by: cyberthirst <cyberthirst.eth@gmail.com>