fix[lang]: allow `print()` schema larger than 32 bytes (#4456)
historically, the schema generated internally by the `print()` builtin
was not allowed to be larger than 32 bytes, ostensibly to simplify the
code for generating the IR, leading to a panic if the user tried to
print a sufficiently complex type (i.e. with an ABI type larger than
32 bytes).

this commit replaces the code for generating the schema in IR with
`Expr._make_bytelike` (along with a small refactor for `_make_bytelike`
to make it a classmethod instead of an instance method), generalizing
to schemas of any size.