fix[venom]: fix sccp resolution of truthy `jnz` (#4505)
this commit fixes SCCP traversal - when a jnz has a truthy condition
that is not equal to 1, it gets transformed to a `jmp`, however, during
lattice construction, both branches are explored and therefore some
phis are suboptimally resolved to BOTTOM, even though they can
theoretically be resolved to a proper value since one of the input bbs
will actually be pruned from the CFG.

this issue is *not* visible in the regular pipeline, since SCCP is run
multiple times and thus the phis get resolved after the second SCCP run.
however, it is visible in tests, and would be visible in the regular
pipeline if SCCP were only run one time.

this commit also removes a `FlowWorkItem` addition, which is redundant
since all the cfg altering instructions are already handled in
`_visit_expr`.