feat[test]: add hevm to `get_contract` harness (#4499)
add `hevm` to the `get_contract` harness. if the `--hevm` flag
is set when running the test suite, the source code argument to
`get_contract()` will additionally be compiled and equivalence-checked
between venom and non-venom bytecode, using `hevm equivalence`.

individual tests can be marked (as is done in this commit) for
inclusion in the hevm CI job, with `pytest.mark.hevm`, and then any
calls to `get_contract` will run the equivalence checker as described
above.

the idea behind this commit is to makes it easy to increase hevm
equivalence coverage in our test suite, simply by manually annotating
individual tests with `pytest.mark.hevm`.