feat[venom]: allow labels to be assigned to variables (#4514)
this commit allows labels to be assigned to variables. it updates
assumptions in `SCCP` and the venom parser which blocked that behavior.

it also updates the text format for `djmp` to be like `jnz` (condition
is first). it also fixes a bug in SimplifyCFG where `merge_jump` did
not rename label operands.

---------

Co-authored-by: Charles Cooper <cooper.charles.m@gmail.com>