feat[venom]: add `calloca` instruction (#4376)
this commit promotes the callee buffer (allocated at call site) to an
alloca-like instruction, `calloca`. this allows us to treat all buffers
generated by the frontend abstractly, instead of needing to do memory
analysis.

`calloca` additionally has a parameter to help identify which call-site
it is associated with, as a hint for alias and reordering analyses.