fix[venom]: fix var defined check for unreachable blocks (#4518)
Fixed an error with `venom_check_ctx` that was triggered with
unreachable basic blocks in the variable definition check (since
unreachable blocks were not visited in the `VarDefinition` analysis).

This commit fixes two root issues: it runs the analysis on all basic
blocks, not just the unreachable ones. It also fixes the lattice
initialization so that it converges correctly no matter the traversal
order.

---------

Co-authored-by: Charles Cooper <cooper.charles.m@gmail.com>