feat[venom]: make `revert` a bb terminator (#4529)
this commit makes `revert` a basic block terminator. previously
`revert` was left as a non-terminating instruction since there could
be the possibility of reordering. however, since it terminates, it is
cleaner to disallow any control flow after the `revert` instruction.

it also fixes a minor bug discovered in `ir_node_to_venom`, where the
translator was appending instructions after a `revert` instruction (it
would have been appending them to a dead block in any case).