fix[lang]!: forbid calling `__default__` (#4371)
This commit forbids `extcall` and `staticcall`s to `__default__`. This
was not possible before the introduction of `__interface__()` and
`__at__()`; however, now that we have the possibility of casting
a module to its interface, it is possible to actually call the
`__default__()` function, which generates a "normal" ABI-encoded
call including the method id for `__default__()`. That could cause a
selector collision with a different function at target, and also leads
to confusing behavior (`__default__` gets called when no selector
matches).

The error message recommends the user to use `raw_call()` for this
use case.

---------

Co-authored-by: cyberthirst <cyberthirst.eth@gmail.com>
Co-authored-by: Charles Cooper <cooper.charles.m@gmail.com>