refactor[venom]: add dom walk property (#4546)
Refactor `DominatorTreeAnalysis` to be more clear regarding cfg walk
usage. Add a dominator post-order traversal property.

The `MakeSSA` pass is refactored to use dom tree walk instead of CFG
post walk. Traversing based on the dominator tree is not technically
mandatory but it aligns with the algorithm's reliance on dominance and
simplifies the renaming phase.