feat[lang]!: move sqrt to new stdlib `math` module (#4520)
this commit moves the `sqrt()` builtin to a new, pure vyper stdlib
implementation.

it includes machinery to handle extension of the stdlib in the
future. the benefits of having machinery to handle code written in pure
vyper in the stdlib is that it will be easier for external contributors
to contribute and audit. it also uses the native import machinery,
rather than the existing ad-hoc mechanism for writing vyper from inside
the compiler, which simplifies the implementation. that mechanism for
writing pure vyper builtins is also removed in this commit (previously,
this machinery had been used specifically to handle "pure vyper"
builtins, the only one of which was `sqrt()`).

---------

Co-authored-by: Charles Cooper <cooper.charles.m@gmail.com>