fix[lang]!: ban calling nonreentrant functions from nonreentrant functions (#4574)
this commit disallows calling nonreentrant functions from other
nonreentrant functions, including nonreentrant functions that are
directly called, or reachable anywhere in the call graph. the reason
for this is that the current codegen has very unintuitive behavior --
since the outer function sets the lock, the inner function will revert
on checking if the lock is set. better to ban the behavior and work on
how to relax it (if at all) later.

---------

Co-authored-by: cyberthirst <cyberthirst.eth@gmail.com>