fix[lang]: only reserve builtins at the top level (#4578)
commit 1591a1229cf837ccca introduced a regression in snekmate, which is
that imports of the form `from .. import math` are rejected, due to the
assertion that `level == 0` on L299 inside of _load_builtin_import.

this was already a latent bug for several versions, but was not
previously tripped, since the only builtin namespace was `ethereum.ercs`
(and users presumably didn't have modules named `ethereum`).

this commit fixes builtin detection by protecting the builtin check with
an additional requirement that `level == 0` (indicating an absolute
import).

misc/refactor:
- add `from_builtin` property on FileInputs that come from the builtin
  loader

---------

Co-authored-by: cyberthirst <cyberthirst.eth@gmail.com>