fix[venom]: fix handling of params in venom_to_assembly
previously, params were handled one at a time. however, steps 4 and 6
could POP or SWAP the params while traversing the param list, which is
invalid since the stack is in an invalid state. for example,

```
function foo {
    %1 = param  ; dead variable
    ; --> POP inserted here! <--
    %2 = param
    %3 = param  ; %3 is scheduled before %2
    ; --> SWAP inserted here! <--
    %4 = param
    mstore %4, %3
    ret %2
}
```

this was previously not visible because of how ir_node_to_venom
generates params; they are very well-behaved (and always live).