refactor[venom]: extract liveness and cfg data structures (#4595)
this commit removes the data structures related to liveness and
the CFG from venom instructions/basic blocks. this makes it so that
the lifetime of the data structures is tied to the lifetime of the
analyses, rather than needing to invalidate on the instructions/basic
blocks, making them more "pure".

it also starts a pattern of deleting attributes from analyses at
invalidation, making it so that if we accidentally try to access
analyses that have been invalidated, an ugly `AttributeError` is
thrown.

misc/refactor:
- clean up some type signatures

---------

Co-authored-by: Hodan <plodeada@fit.cvut.cz>