fix[venom]: fix inliner return items
this fixes the function inliner returning items. currently in vyper,
functions do not return stack items, so there was an assumption that
`ret` had no arguments besides the return pc. this commit adds the store
instructions so that the return items are set properly.

it also fixes a bug in InstUpdater, which is that invoke can have no
outputs.