feat[lang]: nonreentrancy by default (#4563)
to control the default nonreentrancy behavior on a module-by-module
basis. when the pragma is set to `on`, all the external functions and
getters in the file will be nonreentrant (excluding pure functions,
which can never be nonreentrant).

this is a step towards making nonreentrancy the default for the
language while providing a backwards-compatible upgrade path for users
in the 0.4.x series. in the future (probably starting in 0.5.0), the
default will be `pragma nonreentrancy on`, which will have to be opted
out of with the pragma.

---------

Co-authored-by: cyberthirst <cyberthirst.eth@gmail.com>