fix[lang]: disallow duplicate getter annotations (#4623)
state variables with getters can currently compile with
duplicate `public` or `reentrant` annotations, e.g.
`x: public(public(uint256))`. this commit blocks the behavior.

it also adds test cases for this behavior, as well as analogous test
cases for `immutable` and the other location annotations (which are
correctly blocked, but good to have the tests as well).

---------

Co-authored-by: Charles Cooper <cooper.charles.m@gmail.com>