feat[venom]: improve phi elimination pass (#4635)
this commit improves the phi elimination pass (introduced in
6b670e73d8f20a5321) when there are multiple phis in the chain. if the
innermost phi has multiple origins, it would block elimination of outer
phis which use that phi. for example,

```llvm
%a = 1
%b = 2
%c = phi %a, %b  ; has two origins
%d = %c
%e = %d
%f = phi %d, %e
```

prior to this commit, `%f = phi %d, %e` would not reduce, because its
origins were detected as `%a = 1` *and* `%b = 2`. this commit modifies
the graph traversal so that further traversal of `%c`'s origins is
blocked from the perspective of `%f`, allowing `%f` to be reduced to
`%c`.

---------

Co-authored-by: Charles Cooper <cooper.charles.m@gmail.com>