fix[codegen]: fix removal of side effects in concat (#4644)
concat would remove side effects for zero-length arguments. fix by
removing the fastpath.

as the test case shows, this pattern is not common in user-code.

this commit also reverts the optimization introduced in a06f7df76fc07,
and replaces it with a tag on the `IRnode`, since otherwise the
`~empty` intrinsic can show up in the IR at the time we lower to
assembly (which is an error).

---------

Co-authored-by: cyberthirst <cyberthirst.eth@gmail.com>