feat[venom]: implement new calling convention (#4482)
this commit updates the IRnode->venom translator to use
stack arguments when the arguments are word-size and under
a certain threshold. to ensure that all the contracts in
`tests/functional/examples/thirdparty/` compile, the threshold
is set to 6. we could relax this later if we have interprocedural
optimizations which can migrate the stack arguments back to memory.

it also updates the inliner and mem2var to handle cases from the new
calling convention.

---------

Co-authored-by: Charles Cooper <cooper.charles.m@gmail.com>