perf[venom]: improve CSE elimination speed (#4607)
in the CSE elimination pass, avoid `O(n*k)` (where `n` == number of
instructions, and `k` == number of available expressions) copies
by replacing the dict with a HAMT. HAMT is an immutable dict with
`O(1)` copy (copy by reference), and `O(log N)` insert/delete/update
operations, each returning a fresh reference to a new HAMT.

we pull in the external package `immutables`, which is stable, shares
maintainership with CPython, and seems mature (receives bugfixes every
now and again, but mostly not adding new features). we pull in the
external library since HAMTs should be generally useful for the types
of lattice operations we do, and we can probably use them to optimize
other passes as well.

this commit also adds the use of `lru_cache` for some `Flag` operations
(which are bizarrely a perf hotspot)