feat[venom]: disable legacy optimizer in venom pipeline (#4411)
this commit disables the legacy optimizer when the venom pipeline
is enabled. this makes the pipeline faster and removes the "training
wheels" for venom (since we can't depend on optimizations in the legacy
pipeline).

after this commit, we see reductions in instruction count across the
contracts in `tests/functional/examples/thirdparty/`(!).

since there are still some minor differences between having the
legacy optimizer turned on (which amount to tuning differences)
and there could be some benefit to studying these, we add a
`VENOM_ENABLE_LEGACY_OPTIMIZER` env flag to enable it for internal use.