refactor[venom]: don't do load elimination in memmerge pass
the load elimination makes the memmerging pass more complex, and as of
3b0ee7804fda6fb, load elimination is also handled (more correctly) in
the `RemoveUnusedVariables` pass.

this incidentally fixes a bug -- the memmerging pass was removing
`mload`s incorrectly in the case that they reach an `msize` instruction.

after this commit, the removal of `mload` is blocked by the presence of
a reachable `msize` -- see the changed test.