refactor[tool]!: update storage layout format (#4495)
update the storage layout format. previously, the storage
layout was stored in a file in the output bundle named
`"storage_layout.json"`. however, that is problematic for when we
generalize to bundles with multiple compilation targets. this PR adds
a mapping to the manifest which maps from contracts to their respective
storage layout overrides, and changes the storage layout json files to
use the original name from the filesystem, mapping more cleanly back to
the user's filesystem.

this commit additionally refactors `ABIInput` to `JSONInput` so that we
can handle more types of json input files more generally.