feat[lang]: `@raw_return` decorator (#4568)
this commit adds a `@raw_return` decorator which allows a contract
to directly return bytes without the compiler adding ABI-encoding
routines. this is particularly useful for calling proxies or other
helper contracts, where you want to return the output bytes directly to
the user without adding another layer of `(bytes)` encoding.

a `raw_return()` builtin was considered, which would essentially
compile directly to an EVM `return` instruction, but it was rejected
since it can obfuscate control flow (e.g., calling `raw_return()`
in a deeply nested internal function will terminate execution and
commit state changes, which might not be obvious to a caller). the
`@raw_return` decorator was chosen because, a) we can control that it
only gets applied to `@external` functions, and b), so that all paths
in the raw-returning function are treated equally w.r.t encoding of the
return data.

---------

Co-authored-by: cyberthirst <cyberthirst.eth@gmail.com>