fix[parser]: generate unique nodes from python AST singletons (#4677)
this commit fixes an issue where operator nodes (e.g., `+`) would
be annotated with the same source info, resulting in confusing error
messages for exceptions involving operator nodes. this is due to the
python parser parsing all operator nodes as singletons (see note). this
commit addresses the issue by creating a (shallow) copy of the node
prior to annotating it, so all nodes in the AST are distinct objects.

since the new approach generically handles `python_ast.expr_context`,
a prior special-case fix for the Load/Store AST nodes can be removed.

this commit also fuses the location marking with the
`AnnotatingVisitor`, which improves clarity (and also performance,
since the number of passes is reduced from 2 to 1).

references:
- https://docs.python.org/3.13/library/ast.html#ast.AsyncWith

  > When a string is parsed by `ast.parse()`, operator nodes
  (subclasses of `ast.operator`, `ast.unaryop`, `ast.cmpop`,
  `ast.boolop` and `ast.expr_context`) on the returned tree will be
  singletons. Changes to one will be reflected in all other occurrences
  of the same value (e.g. `ast.Add`).

---------

Co-authored-by: Charles Cooper <cooper.charles.m@gmail.com>