feat[test]: add more ctor abi decoder tests (#4648)
abi-decoding of constructor arguments lacked test coverage for edge
cases tested in runtime decoding of memory. this commit reuses the
existing abi-decode tests and adds machinery to reuse the payloads as
calldata to `__init__()`, increasing test coverage of ctor abi-decoding.

note that since the semantics of the constructor decoder are different
as for calldata and code, the pointers allowed to point to OOB
locations. hence, there is some divergence where the existing tests
(where the payload is in memory) test for failure, but the ctor test
decodes a valid payload.

semantics of some of the existing abi-decoding tests were intentionally
refactored to achieve better readability as some of the test constructs
weren't necessary to achieve the test's purpose.

further, the commit adds tests for various edge case scenarios in the
constructor context like reading immutables before they were assigned,
initializing immutables with constants, or assigning the same immutables
in a loop.

---------

Co-authored-by: Charles Cooper <cooper.charles.m@gmail.com>