feat[tool]: Add extra debugging output
This commit contains two main changes.

The first are two new output modes: the `symbol_map` and
`symbol_map_runtime`. These contain both contain the raw "symbol map"
produced during the respective assembly of the deployment and runtime
code into bytecode. This information is crucial for determining the
exact starting position of internal functions in the assembled bytecode.

The second component is a change to the existing metadata output. When
the user selects the experimental codegen, the metadata for internal
functions will contain a list of the parameter names which are passed via the
stack in `venom_via_stack` key. Similarly, if the function returns its
value via the stack, the "venom_return_via_stack" key records this
as a boolean. These keys are omitted if experimental codegen is not
selected. (Credit where credit is due: this feature was primarily
developed by Charles Coop).

Co-Authored-By: Charles Cooper <cooper.charles.m@gmail.com>