refactor[codegen]: refactor assembly (#4663)
this commit is a follow-on to 90c952c06ac31a. it removes recursion from
the assembler. previously, the assembly was structured with the runtime
code as a "subassembly", which is recursively compiled during lowering
to bytecode.

this commit instead pre-compiles the runtime code into a data blob in
the initcode before entering the assembler, decreasing the conceptual
overhead of the assembler. this can be seen by the simplification of
`venom_to_assembly.py`, for instance.

to handle some of the assembly-time constant folding specifically
related to initcode memory allocation, it also adds the CONSTREF data
structure, which represents constants which are inserted into the
assembly.

lastly, the main routine for compiling (legacy) IR to assembly has been
wrapped in an object, so that the recursion does not need to keep track
of so much state.

---------

Co-authored-by: Harry Kalogirou <harkal@nlogn.eu>