feat[lang]: add support for single line multi-imports (#4762)
this commit refactors the way imports are encoded to allow for importing
multiple members from the same module. this is done by following more
closely python's ast. in particular the new `alias` node is
uncapitalized because so is the case in python.

python allows (but discourages, see PEP-8) multiple module imports on
the same line. we decided not to support this as we would discourage it
anyways. a validator is thus added which enforces this decision. to lift
this restriction, removing the validator is enough

moving the `"." in alias` check from `_handle_Import` to `_add_imports`
does not change execution, even when execution comes from
`_handle_ImportFrom`, because the parser forbids periods in aliases