fix[lang]: disallow events from having modules as members (#4768)
this commit fixes the compiler panicking on events with modules as
member. this is done by moving the struct check up to all user-defined
types. at the moment, `_as_tuple_member` is true for all types except
modules, fulfilling our requirement.

this commit also makes the crash message on `abi_type` not being
implemented clearer. this message should never be seen by users, so that
change is only useful for compiler developers.