fix[lang]: fix incorrect handling of tuples in nested subscripts (#4776)
this commits adds some logic to gracefully handle typechecking of nested
subscripts in the presence of tuple types, rather than panicking. it
assumes the only valid indices for tuples are constant integers, this is
guaranteed by the parser.

it also adds a new test that checks these kinds of nested subscripts.

---------

Co-authored-by: Charles Cooper <cooper.charles.m@gmail.com>