fix[venom]: fix `fix_mem_loc` (#4789)
fix the `fix_mem_loc()` util.

this case is wrong in the case that the op is an `IRAbstractMemLoc`,
because `.value` is just its id, not any meaning related to its pointer
offset.

previously this case (i.e., where the item is an `IRAbstractMemLoc`)
didn't get triggered because at that stage of compilation, all
`IRAbstractMemLoc`s are only contained in `*alloca` instructions. this
fixes the case for follow on PRs.

it also fixes a case in `may_overlap()` which incorrectly returns false
when comparing concrete memory with abstract memory. this is currently
unreachable in the codebase during alias analysis since we have an
invariant that we do not simultaneously have concrete and abstract
memory locations in generated venom code, but is not true for more
general venom code.