fix[venom]: fix clobber logic in dse (#4785)
this commit fixes overly aggressive (incorrect) clobbers in DSE.

the root issue is that the memory semantics of calls were modeled
incorrectly -- the `size` argument was being passed as the size to
`MemoryLocation`, but in the EVM, the size is just the maximum number of
bytes that is allowed to be written at that location, the callee
(invoked by `*CALL`) can write fewer bytes at runtime. the import to us
is that this results in the previous write to that memory buffer not
necessarily being clobbered.

this fixes the issue by modeling the size written as `None`, so that
passes don't analyze the buffer as being clobbered. this results
conversely in overly conservative passes now since the size is
technically modeled as unlimited (but we have a bound from the EVM
semantics); this will be fixed at a later date.

---------

Co-authored-by: Hodan <plodeada@fit.cvut.cz>